--- rest-server/bin/harness-daemon
@@ -1,49 +1,0 @@
-#!/usr/bin/env bash
-
-. harness-env
-
-PIDFILE=$1
-
-shift
-
-export HARNESS_REST_PORT=9090
-
-case ${1} in
-    -p|--port)
-        export HARNESS_REST_PORT="${2}"
-        echo " > HARNESS_REST_PORT: ${HARNESS_REST_PORT}"
-        ;;
-
-    *)
-        export CONFIG=$1
-        case ${2} in
-            -p|--port)
-                export HARNESS_REST_PORT="${3}"
-                echo " >> HARNESS_REST_PORT: ${HARNESS_REST_PORT}"
-                ;;
-
-            *) ;;
-        esac
-    ;;
-esac
-
-if [ -z ${CONFIG} ] ; then
-    export CONFIG="${HARNESS_HOME}/conf/config.conf"
-fi
-
-if [ ! -f ${CONFIG} ]; then
-    echo -e "${Red}Config file '${CONFIG}' not found!${Reset}"
-    exit 1
-fi
-
-if [ ! -f "${HARNESS_HOME}/bin/main" ]; then
-    echo -e "${Red}Main file '${HARNESS_HOME}/bin/main' not found!${Reset}"
-    exit 1
-fi
-
-echo "Pid file: ${PIDFILE}"
-echo -e "${Green}${HARNESS_HOME}/bin/main -Dconfig.file=${CONFIG} -Drest-server.port=${HARNESS_REST_PORT}${Reset}"
-
-exec ${HARNESS_HOME}/bin/main -Dconfig.file=${CONFIG} -Drest-server.port=${HARNESS_REST_PORT} <&- > /dev/null 2>&1 &
-echo $! > ${PIDFILE}
-cat ${PIDFILE}
