--- rest-server/bin/harness-engine
@@ -1,22 +1,0 @@
-#!/usr/bin/env bash
-
-. header
-
-COMMAND=$1
-shift
-
-case ${COMMAND} in
-
-    new)
-        ${HARNESS_HOME}/bin/harness-engine-new $1 $2 $3 ;;
-
-    update)
-        ${HARNESS_HOME}/bin/harness-engine-update $1 $2 $3 ;;
-
-    delete)
-        ${HARNESS_HOME}/bin/harness-engine-delete $1 ;;
-
-    *)
-        echo -e "${Red}Unknown command: '$COMMAND' for 'engine' section. Need use 'new', 'update' or 'delete'${Reset}"
-        exit 1 ;;
-esac
