+++ rest-server/bin/harness-engine-delete
@@ -1,12 +1,31 @@
 #!/usr/bin/env bash
 
-. harness-env
+. "${HARNESS_HOME}/bin/harness-env"
 
-RESOURCE_ID=$1
+case "$1" in
+    -c) ENGINE_CONFIG=$2
+        shift ;;
+    *) RESOURCE_ID=$1 ;;
+esac
 
-if [ -z "$RESOURCE_ID" ]; then
-    echo -e "${Red}Engine id not found!${Reset} Expect 'harness engine delete <engine-id>'"
+if [ -z "$ENGINE_CONFIG" ] && [ -z "$RESOURCE_ID" ]; then
+    echo -e "${RED}Engine config file or resource id not specified!${NC}"
+    echo -e "Expect command: ${CYAN}harness delete [-c <some-engine.json> | <some-resource-id]${NC}"
     exit 1
 fi
 
-${AML_HOME}/bin/engine.py delete ${RESOURCE_ID}
+if [ ! -f ${ENGINE_CONFIG} ]; then
+    echo -e "${RED}Engine config file [${ENGINE_CONFIG}] not found!${NC}"
+    echo -e "Expect command: ${CYAN}harness delete [-c <some-engine.json> | <some-resource-id]${NC}"
+    echo -e "Verify that the correct path is specified."
+    exit 1
+fi
+
+PYTHON_ARGS="delete"
+if [ -n "$RESOURCE_ID" ]; then PYTHON_ARGS="${PYTHON_ARGS} ${RESOURCE_ID}"; fi
+if [ -n "$ENGINE_CONFIG" ]; then PYTHON_ARGS="${PYTHON_ARGS} -c ${ENGINE_CONFIG}"; fi
+
+
+echo -e "${CYAN}Run ${HARNESS_HOME}/bin/engine.py ${PYTHON_ARGS}${NC}"
+
+${HARNESS_HOME}/bin/engine.py ${PYTHON_ARGS}
