+++ rest-server/bin/harness-engine-train
@@ -1,0 +1,21 @@
+#!/usr/bin/env bash
+
+. "${HARNESS_HOME}/bin/harness-env"
+
+ENGINE_CONFIG=$1
+shift
+
+if [ -z "$ENGINE_CONFIG" ]; then
+    echo -e "${RED}Engine config file not specified!${NC}"
+    echo -e "Expect command: ${CYAN}harness train <some-engine.json>${NC}"
+    exit 1
+fi
+
+if [ ! -f ${ENGINE_CONFIG} ]; then
+    echo -e "${RED}Engine config file [${ENGINE_CONFIG}] not found!${NC}"
+    echo -e "Expect command: ${CYAN}harness train <some-engine.json>${NC}"
+    echo -e "Verify that the correct path is specified."
+    exit 1
+fi
+
+${HARNESS_HOME}/bin/command.py train ${ENGINE_CONFIG}
