+++ rest-server/bin/harness-env
@@ -1,17 +1,27 @@
 #!/usr/bin/env bash
 
-# Figure out where Harness is installed
-export HARNESS_HOME="$(cd `dirname $0`/..; pwd)"
+export REST_SERVER_HOST=0.0.0.0
+export REST_SERVER_PORT=9090
+
+export MONGO_HOST=0.0.0.0
+export MONGO_PORT=27017
+
+export HARNESS_LOG_CONFIG="${HARNESS_HOME}/conf/logback.xml"
+export HARNESS_LOG_PATH="${HARNESS_HOME}/logs"
 
 # Reset
-Reset='\033[0m'       # Text Reset
+NC='\033[0m'           # Text Reset
 
 # Regular Colors
-Black='\033[0;30m'        # Black
-Red='\033[0;31m'          # Red
-Green='\033[0;32m'        # Green
-Yellow='\033[0;33m'       # Yellow
-Blue='\033[0;34m'         # Blue
-Purple='\033[0;35m'       # Purple
-Cyan='\033[0;36m'         # Cyan
-White='\033[0;37m'        # White
+RED='\033[0;31m'          # Red
+GREEN='\033[0;32m'        # Green
+YELLOW='\033[0;33m'       # Yellow
+BLUE='\033[0;34m'         # Blue
+PURPLE='\033[0;35m'       # Purple
+CYAN='\033[0;36m'         # Cyan
+WHITE='\033[0;37m'        # White
+
+LINE="=================================================================="
+
+GLINE="${GREEN}${LINE}"
+RLINE="${LINE}${NC}"
