+++ rest-server/bin/harness-status
@@ -1,5 +1,15 @@
 #!/usr/bin/env bash
 
-. harness-env
+. "${HARNESS_HOME}/bin/harness-env"
+
+echo -e "${GLINE}"
+echo "HARNESS_HOME........${HARNESS_HOME}"
+echo "REST_SERVER_HOST....${REST_SERVER_HOST}"
+echo "REST_SERVER_PORT....${REST_SERVER_PORT}"
+echo "MONGO_HOST..........${MONGO_HOST}"
+echo "MONGO_PORT..........${MONGO_PORT}"
+echo "HARNESS_LOG_CONFIG..${HARNESS_LOG_CONFIG}"
+echo "HARNESS_LOG_PATH....${HARNESS_LOG_PATH}"
+echo -e "${RLINE}"
 
 ${HARNESS_HOME}/bin/command.py status
