--- rest-server/bin/harness-train
@@ -1,12 +1,0 @@
-#!/usr/bin/env bash
-
-. harness-env
-
-RESOURCE_ID=$1
-
-if [ -z "$RESOURCE_ID" ]; then
-    echo -e "${Red}Engine id not found!${Reset} Expect 'harness train <engine-id>'"
-    exit 1
-fi
-
-${HARNESS_HOME}/bin/command.py train ${RESOURCE_ID}
