--- rest-server/bin/header
@@ -1,17 +1,0 @@
-#!/usr/bin/env bash
-
-# Figure out where Harness is installed
-export HARNESS_HOME="$(cd `dirname $0`/..; pwd)"
-
-# Reset
-Reset='\033[0m'       # Text Reset
-
-# Regular Colors
-Black='\033[0;30m'        # Black
-Red='\033[0;31m'          # Red
-Green='\033[0;32m'        # Green
-Yellow='\033[0;33m'       # Yellow
-Blue='\033[0;34m'         # Blue
-Purple='\033[0;35m'       # Purple
-Cyan='\033[0;36m'         # Cyan
-White='\033[0;37m'        # White
